Start BCD work
This implements AF on existing operations where required and adds
facilities for binary-coded decimal adjustment on addition.

The flag setters were modified, with the setf code now using the
defsetf macro directly. This freed up the set-flag symbol to be used
for a function to set a flag to 1 (t).